refactor: Regroup icons related components inside own folder
