+++ src/modules/search/components/SuggestionItem.jsx
@@ -5,8 +5,8 @@ import ListItemIcon from 'cozy-ui/transpiled/react/ListItemIcon'
 import ListItemText from 'cozy-ui/transpiled/react/ListItemText'
 
 import { SHARED_DRIVES_DIR_ID } from 'constants/config'
-import FileIconMime from 'modules/filelist/FileIconMime'
-import FileIconShortcut from 'modules/filelist/FileIconShortcut'
+import FileIconMime from 'modules/filelist/icons/FileIconMime'
+import FileIconShortcut from 'modules/filelist/icons/FileIconShortcut'
 import SuggestionItemTextHighlighted from 'modules/search/components/SuggestionItemTextHighlighted'
 import SuggestionItemTextSecondary from 'modules/search/components/SuggestionItemTextSecondary'
